+++ account_banking_fr_lcr/README.rst
@@ -29,7 +29,7 @@ select the LCR payment mode.
 
 .. image:: https://odoo-community.org/website/image/ir.attachment/5784_f2813bd/datas
    :alt: Try me on Runbot
-   :target: https://runbot.odoo-community.org/runbot/121/9.0
+   :target: https://runbot.odoo-community.org/runbot/121/10.0
 
 Bug Tracker
 ===========
@@ -50,9 +50,9 @@ Contributors
 Maintainer
 ----------
 
-.. image:: http://odoo-community.org/logo.png
+.. image:: https://odoo-community.org/logo.png
    :alt: Odoo Community Association
-   :target: http://odoo-community.org
+   :target: https://odoo-community.org
 
 This module is maintained by the OCA.
