+++ account_banking_fr_lcr/data/account_payment_method.xml
@@ -1,6 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
-<odoo>
-<data noupdate="1">
+<odoo noupdate="1">
 
 
 <record id="fr_lcr" model="account.payment.method">
@@ -11,5 +10,4 @@
 </record>
 
 
-</data>
 </odoo>
