+++ account_banking_fr_lcr/demo/lcr_demo.xml
@@ -1,6 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
-<odoo>
-<data noupdate="1">
+<odoo noupdate="1">
 
 
 <record id="payment_mode_lcr" model="account.payment.mode">
@@ -12,5 +11,4 @@
 </record>
 
 
-</data>
 </odoo>
