+++ account_banking_fr_lcr/models/account_payment_order.py
@@ -2,8 +2,8 @@
 # © 2014-2016 Akretion (Alexis de Lattre <alexis.delattre@akretion.com>)
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
 
-from openerp import models, fields, api, _
-from openerp.exceptions import UserError
+from odoo import models, fields, api, _
+from odoo.exceptions import UserError
 from unidecode import unidecode
 
 LCR_DATE_FORMAT = '%d%m%y'
