--- l10n_fr_siret/partner_demo.xml
@@ -1,17 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<!--
-    Copyright (C) 2014 Akretion (http://www.akretion.com/)
-    @author: Alexis de Lattre <alexis.delattre@akretion.com>
-    The licence is in the file __openerp__.py
--->
-
-<openerp>
-    <data noupdate="1">
-
-        <record id="base.res_partner_12" model="res.partner">  <!-- C2C -->
-            <field name="siren">433698578</field>
-            <field name="nic">00039</field>
-            <field name="company_registry">Chambery</field>
-        </record>
-    </data>
-</openerp>
